requirements.txt
feedparser
requests
groq
schedule
pytz
python-dotenv
playwright
